moves timeout tests to an integration test
